requirements.txt
blinker==1.8.2
click==8.1.7
email_validator==2.1.1
Flask==3.0.3
Flask-WTF==1.2.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
PyYAML==6.0.1
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.17.0
markupsafe==2.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2